+++ src/tabular_energy.cpp
@@ -50,7 +50,7 @@ TabularEnergy::TabularEnergy(const ACE& ace, size_t i, size_t JED)
 
   // Read tables
   for (uint32_t j = 0; j < NE; j++) {
-    uint32_t loc = JED + ace.xss<uint32_t>(i + 2 + 2 * NR + NE + j) - 1;
+    uint32_t loc = static_cast<uint32_t>(JED) + ace.xss<uint32_t>(i + 2 + 2 * NR + NE + j) - 1;
     tables_.emplace_back(ace, loc);
   }
 }
